# Improve performance of Delete algorithm