## Description


Improve delete performance by re-inserting whole subtrees that become under-populated
rather than re-inserting leaves one by one.

## Check List

Have you:

- Added unit tests? N/A, relevant tests already existed.

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

#208 

## Benchmark Results

~20% delete performance improvement.

```
name                                          old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4       2.55µs ±13%    2.73µs ±16%   +6.90%  (p=0.017 n=12+12)
IntersectsLineStringWithLineString/n=100-4      41.8µs ±13%    40.9µs ± 5%     ~     (p=0.687 n=13+13)
IntersectsLineStringWithLineString/n=1000-4      598µs ± 8%     591µs ± 8%     ~     (p=0.719 n=15+12)
IntersectsLineStringWithLineString/n=10000-4    10.5ms ±30%     9.5ms ± 8%   -9.27%  (p=0.011 n=14+14)
LineStringIsSimpleZigZag/10-4                   2.22µs ±23%    2.25µs ±11%     ~     (p=0.254 n=13+15)
LineStringIsSimpleZigZag/100-4                  70.0µs ±33%    66.1µs ±11%     ~     (p=0.650 n=14+13)
LineStringIsSimpleZigZag/1000-4                 1.10ms ±27%    1.07ms ±15%     ~     (p=0.786 n=13+15)
LineStringIsSimpleZigZag/10000-4                14.3ms ± 8%    14.5ms ± 8%     ~     (p=0.793 n=14+13)
PolygonSingleRingValidation/n=10-4              2.55µs ± 9%    2.60µs ±22%     ~     (p=0.915 n=15+12)
PolygonSingleRingValidation/n=100-4             59.8µs ± 3%    58.5µs ± 5%   -2.19%  (p=0.043 n=13+14)
PolygonSingleRingValidation/n=1000-4            1.10ms ±11%    1.10ms ±13%     ~     (p=0.652 n=15+14)
PolygonSingleRingValidation/n=10000-4           15.4ms ± 8%    14.6ms ± 5%   -4.75%  (p=0.028 n=15+12)
PolygonMultipleRingsValidation/n=4-4            11.0µs ± 7%    11.0µs ± 3%     ~     (p=0.234 n=14+13)
PolygonMultipleRingsValidation/n=36-4            104µs ±10%     102µs ±11%     ~     (p=0.982 n=14+14)
PolygonMultipleRingsValidation/n=400-4          1.26ms ± 2%    1.36ms ±18%   +8.19%  (p=0.000 n=12+14)
PolygonMultipleRingsValidation/n=4096-4         15.5ms ±10%    15.8ms ± 9%     ~     (p=0.297 n=12+14)
PolygonZigZagRingsValidation/n=10-4             25.5µs ± 9%    25.4µs ±10%     ~     (p=1.000 n=15+15)
PolygonZigZagRingsValidation/n=100-4             352µs ± 5%     353µs ± 4%     ~     (p=1.000 n=13+13)
PolygonZigZagRingsValidation/n=1000-4           4.85ms ±19%    4.69ms ±13%     ~     (p=0.155 n=13+14)
PolygonZigZagRingsValidation/n=10000-4          66.1ms ±12%    64.9ms ±12%     ~     (p=0.239 n=14+13)
MultipolygonValidation/n=1-4                     368ns ±19%     363ns ±22%     ~     (p=0.694 n=14+13)
MultipolygonValidation/n=4-4                     918ns ± 8%    1017ns ±30%  +10.74%  (p=0.017 n=15+14)
MultipolygonValidation/n=16-4                   5.93µs ±15%    5.96µs ± 8%     ~     (p=0.545 n=13+13)
MultipolygonValidation/n=64-4                   34.3µs ± 5%    35.6µs ± 9%   +3.90%  (p=0.001 n=14+14)
MultipolygonValidation/n=256-4                   207µs ±11%     219µs ± 7%   +5.68%  (p=0.001 n=14+14)
MultipolygonValidation/n=1024-4                 1.04ms ±10%    1.08ms ±10%     ~     (p=0.061 n=13+14)
MultiPolygonTwoCircles/n=10-4                   7.69µs ±13%    7.54µs ±13%     ~     (p=0.479 n=13+13)
MultiPolygonTwoCircles/n=100-4                   113µs ±11%     112µs ± 7%     ~     (p=0.683 n=15+14)
MultiPolygonTwoCircles/n=1000-4                 1.56ms ±13%    1.60ms ±18%     ~     (p=0.650 n=13+15)
MultiPolygonTwoCircles/n=10000-4                24.7ms ±12%    24.2ms ±10%     ~     (p=0.217 n=14+15)
MultiPolygonMultipleTouchingPoints/n=1-4        6.15µs ± 5%    6.34µs ± 7%     ~     (p=0.051 n=15+14)
MultiPolygonMultipleTouchingPoints/n=10-4       53.4µs ± 9%    53.3µs ± 9%     ~     (p=0.786 n=15+13)
MultiPolygonMultipleTouchingPoints/n=100-4       608µs ± 8%     608µs ± 7%     ~     (p=0.793 n=14+13)
MultiPolygonMultipleTouchingPoints/n=1000-4     8.13ms ±10%    8.11ms ± 8%     ~     (p=0.920 n=13+13)
MarshalWKB/polygon/n=10-4                        209ns ±11%     201ns ±22%     ~     (p=0.089 n=12+13)
MarshalWKB/polygon/n=100-4                       639ns ±42%     655ns ±42%     ~     (p=0.639 n=12+14)
MarshalWKB/polygon/n=1000-4                    5.29µs ±133%   5.88µs ±160%     ~     (p=1.000 n=13+13)
MarshalWKB/polygon/n=10000-4                    35.1µs ±92%   53.2µs ±146%     ~     (p=0.141 n=13+14)
UnmarshalWKB/polygon/n=10-4                      351ns ±24%     350ns ±11%     ~     (p=0.494 n=13+12)
UnmarshalWKB/polygon/n=100-4                     837ns ±40%     714ns ±11%  -14.78%  (p=0.011 n=14+12)
UnmarshalWKB/polygon/n=1000-4                   8.68µs ±60%    7.41µs ±78%     ~     (p=0.402 n=14+13)
UnmarshalWKB/polygon/n=10000-4                  40.6µs ±51%    44.8µs ±49%     ~     (p=0.185 n=14+13)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                             50.1µs ± 6%    50.9µs ±11%     ~     (p=0.462 n=12+14)
Intersection/n=100-4                            98.8µs ± 9%   101.8µs ±17%     ~     (p=0.311 n=13+13)
Intersection/n=1000-4                            420µs ±13%     436µs ±19%     ~     (p=0.354 n=15+14)
Intersection/n=10000-4                          3.73ms ±26%    3.67ms ±11%     ~     (p=0.983 n=15+14)
NoOp/n=10-4                                     4.03µs ± 9%    3.96µs ±12%     ~     (p=0.312 n=14+14)
NoOp/n=100-4                                    13.2µs ± 8%    13.1µs ±11%     ~     (p=0.652 n=14+15)
NoOp/n=1000-4                                   95.0µs ±13%    95.2µs ±15%     ~     (p=0.983 n=15+14)
NoOp/n=10000-4                                   946µs ±17%    1027µs ±16%   +8.57%  (p=0.035 n=14+14)
pkg:github.com/peterstace/simplefeatures/rtree goos:linux goarch:amd64
Delete/n=100-4                                  56.7µs ±10%    41.7µs ± 9%  -26.49%  (p=0.000 n=15+15)
Delete/n=1000-4                                  980µs ± 8%     831µs ± 4%  -15.24%  (p=0.000 n=15+14)
Delete/n=10000-4                                57.1ms ±11%    46.1ms ±14%  -19.29%  (p=0.000 n=15+15)

name                                          old alloc/op   new alloc/op   delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4       2.82kB ± 0%    2.82kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=100-4      34.4kB ± 0%    34.4kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=1000-4      237kB ± 0%     237kB ± 0%     ~     (p=1.007 n=14+15)
IntersectsLineStringWithLineString/n=10000-4    3.15MB ± 0%    3.15MB ± 0%     ~     (p=0.507 n=15+14)
LineStringIsSimpleZigZag/10-4                   1.44kB ± 0%    1.44kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/100-4                  21.6kB ± 0%    21.6kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/1000-4                  230kB ± 0%     230kB ± 0%     ~     (p=0.929 n=15+14)
LineStringIsSimpleZigZag/10000-4                2.30MB ± 0%    2.30MB ± 0%     ~     (p=0.992 n=14+15)
PolygonSingleRingValidation/n=10-4              1.47kB ± 0%    1.47kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=100-4             16.2kB ± 0%    16.2kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=1000-4             217kB ± 0%     217kB ± 0%     ~     (p=1.000 n=15+15)
PolygonSingleRingValidation/n=10000-4           2.23MB ± 0%    2.23MB ± 0%     ~     (p=0.171 n=15+15)
PolygonMultipleRingsValidation/n=4-4            5.28kB ± 0%    5.28kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=36-4           43.3kB ± 0%    43.3kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=400-4           476kB ± 0%     476kB ± 0%     ~     (p=0.960 n=15+15)
PolygonMultipleRingsValidation/n=4096-4         4.88MB ± 0%    4.88MB ± 0%   -0.00%  (p=0.026 n=15+14)
PolygonZigZagRingsValidation/n=10-4             12.3kB ± 0%    12.3kB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=100-4             143kB ± 0%     143kB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=1000-4           1.11MB ± 0%    1.11MB ± 0%     ~     (p=0.440 n=13+12)
PolygonZigZagRingsValidation/n=10000-4          13.5MB ± 0%    13.5MB ± 0%     ~     (p=0.433 n=15+15)
MultipolygonValidation/n=1-4                      385B ± 0%      385B ± 0%     ~     (all equal)
MultipolygonValidation/n=4-4                      676B ± 0%      676B ± 0%     ~     (all equal)
MultipolygonValidation/n=16-4                   3.86kB ± 0%    3.86kB ± 0%     ~     (all equal)
MultipolygonValidation/n=64-4                   15.4kB ± 0%    15.4kB ± 0%     ~     (all equal)
MultipolygonValidation/n=256-4                  63.1kB ± 0%    63.1kB ± 0%     ~     (all equal)
MultipolygonValidation/n=1024-4                  259kB ± 0%     259kB ± 0%   -0.00%  (p=0.012 n=14+13)
MultiPolygonTwoCircles/n=10-4                   5.73kB ± 0%    5.73kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=100-4                  62.9kB ± 0%    62.9kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=1000-4                  410kB ± 0%     410kB ± 0%     ~     (p=0.093 n=12+12)
MultiPolygonTwoCircles/n=10000-4                5.65MB ± 0%    5.65MB ± 0%     ~     (p=0.384 n=12+15)
MultiPolygonMultipleTouchingPoints/n=1-4        4.18kB ± 0%    4.18kB ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=10-4       24.3kB ± 0%    24.3kB ± 0%     ~     (p=0.815 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100-4       188kB ± 0%     188kB ± 0%     ~     (p=0.056 n=14+13)
MultiPolygonMultipleTouchingPoints/n=1000-4     2.28MB ± 0%    2.28MB ± 0%     ~     (p=0.629 n=14+15)
MarshalWKB/polygon/n=10-4                         232B ± 0%      232B ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100-4                      1.83kB ± 0%    1.83kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000-4                     16.4kB ± 0%    16.4kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000-4                     164kB ± 0%     164kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10-4                       282B ± 0%      282B ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100-4                    1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000-4                   16.5kB ± 0%    16.5kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000-4                   164kB ± 0%     164kB ± 0%     ~     (all equal)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                             1.19kB ± 0%    1.19kB ± 0%     ~     (all equal)
Intersection/n=100-4                            6.33kB ± 0%    6.33kB ± 0%     ~     (all equal)
Intersection/n=1000-4                           55.0kB ± 0%    55.0kB ± 0%     ~     (p=0.056 n=15+12)
Intersection/n=10000-4                           557kB ± 0%     557kB ± 0%   +0.00%  (p=0.023 n=13+15)
NoOp/n=10-4                                       864B ± 0%      864B ± 0%     ~     (all equal)
NoOp/n=100-4                                    5.68kB ± 0%    5.68kB ± 0%     ~     (all equal)
NoOp/n=1000-4                                   49.5kB ± 0%    49.5kB ± 0%     ~     (all equal)
NoOp/n=10000-4                                   492kB ± 0%     492kB ± 0%     ~     (p=0.061 n=15+14)
pkg:github.com/peterstace/simplefeatures/rtree goos:linux goarch:amd64
Delete/n=100-4                                  4.72kB ± 0%    0.71kB ± 0%  -84.92%  (p=0.000 n=15+15)
Delete/n=1000-4                                 55.1kB ± 0%    24.2kB ± 0%  -56.11%  (p=0.000 n=15+15)
Delete/n=10000-4                                1.52MB ± 0%    0.46MB ± 0%  -69.45%  (p=0.000 n=12+15)

name                                          old allocs/op  new allocs/op  delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4         18.0 ± 0%      18.0 ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=100-4         166 ± 0%       166 ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=1000-4      1.11k ± 0%     1.11k ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=10000-4     17.8k ± 0%     17.8k ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10-4                     5.00 ± 0%      5.00 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/100-4                    75.0 ± 0%      75.0 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/1000-4                    797 ± 0%       797 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10000-4                 8.00k ± 0%     8.00k ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10-4                6.00 ± 0%      6.00 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=100-4               57.0 ± 0%      57.0 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=1000-4               755 ± 0%       755 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10000-4            7.73k ± 0%     7.73k ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4-4              25.0 ± 0%      25.0 ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=36-4              203 ± 0%       203 ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=400-4           2.23k ± 0%     2.23k ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4096-4          22.8k ± 0%     22.8k ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=10-4               64.0 ± 0%      64.0 ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=100-4               654 ± 0%       654 ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=1000-4            4.95k ± 0%     4.95k ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=10000-4           69.5k ± 0%     69.5k ± 0%   -0.00%  (p=0.026 n=15+13)
MultipolygonValidation/n=1-4                      5.00 ± 0%      5.00 ± 0%     ~     (all equal)
MultipolygonValidation/n=4-4                      8.00 ± 0%      8.00 ± 0%     ~     (all equal)
MultipolygonValidation/n=16-4                     27.0 ± 0%      27.0 ± 0%     ~     (all equal)
MultipolygonValidation/n=64-4                     99.0 ± 0%      99.0 ± 0%     ~     (all equal)
MultipolygonValidation/n=256-4                     392 ± 0%       392 ± 0%     ~     (all equal)
MultipolygonValidation/n=1024-4                  1.58k ± 0%     1.58k ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=10-4                     42.0 ± 0%      42.0 ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=100-4                     338 ± 0%       338 ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=1000-4                  2.23k ± 0%     2.23k ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=10000-4                 35.5k ± 0%     35.5k ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=1-4          51.0 ± 0%      51.0 ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=10-4          334 ± 0%       334 ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=100-4       2.99k ± 0%     2.99k ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=1000-4      32.3k ± 0%     32.3k ± 0%     ~     (p=0.629 n=13+15)
MarshalWKB/polygon/n=10-4                         6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100-4                        6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000-4                       6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000-4                      6.00 ± 0%      6.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10-4                       7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100-4                      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000-4                     7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000-4                    7.00 ± 0%      7.00 ± 0%     ~     (all equal)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                               31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=100-4                              31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=1000-4                             31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=10000-4                            31.0 ± 0%      31.0 ± 0%     ~     (all equal)
NoOp/n=10-4                                       21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=100-4                                      21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=1000-4                                     21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=10000-4                                    21.0 ± 0%      21.0 ± 0%     ~     (all equal)
pkg:github.com/peterstace/simplefeatures/rtree goos:linux goarch:amd64
Delete/n=100-4                                    72.0 ± 0%      65.0 ± 0%   -9.72%  (p=0.000 n=15+15)
Delete/n=1000-4                                    509 ± 0%       463 ± 0%   -9.04%  (p=0.000 n=15+15)
Delete/n=10000-4                                 10.5k ± 0%      8.0k ± 0%  -24.33%  (p=0.000 n=14+15)
```